"2:55pm.
```
    inl sum =
        unary_bind <| inl x ->
            inl out = k.block.uter (+) (primal x) |> dr
            {
            out
            bck=inl _ ->
                inl er = k.block.uter (+) (adjoint out)
                add_adjoint x (inl item -> get_adjoint item out)
            }
```

Here is the `sum`. Actually, I am kind of drawing a blank of how one would do a template for this.

The kind of reduction one can backprop is basically just `sum`.

Well, I did backprop through `max` back in late 2015 while I was working on sparse autoencoders, but that was kind of a special thing that won't be repeated.

```
    // TODO: Those chained views might not work. Redesign the tensor so that .view is the default mode.
    inl link_adjoint_view b x = link_adjoint b {x with from=Struct.map (inl x a b -> x .view a .view b) self}
```

Let me do it like this.

...No, actually, no. The default should stay what it is. Things would get difficult otherwise. Actually, I think what I have above should work. It is quite nice.

3:20pm. I really need a breather. Let me add `iter`, remove an argument from `init`, redo the functions in `Learning` that use stuff from `CudaAD`. This should be enough to let me stop for a bit before I move on to testing."