The library is operational. It seems I did a good job.
I suppose I should start work on `init_seq`, but to be honest, I do not really feel like starting a new thing at this point. My mind is dull and I've been awake for 10h at this point. It is going to be hard getting into the zone.

I'll do it tomorrow. Since I do not have to do things like figure out what the right architecture is anymore, I will have the time to do it all in the morning. I'll ignore posting in the PL thread until I am done with work for the day tomorrow.

What did I even do this month? Tree views and the optimized RNNs.

The things in between are a bit blurry for me. I'll refresh my memory by going through the commits and then break off. Adios.