"12:45pm.
```
        inl l2_norm_body size x =
            inl mean x = sum x / to float size
            inm centered = x - mean x
            inm std = sqr centered |> mean |> sqrt
            succ (centered, std)

        inl layer_norm epsilon size x =
            inm x, std = l2_norm_body size x
            x / (std + epsilon)

        inl weight_norm size x =
            inm x, std = l2_norm_body size x
            if_ {cond=std < one; true=x; false=x / std}
```

I am going to have to make `if_` as a primitve. Also `sqr` and `sqrt`.

`if_` is a bit different than the others so I am going to have to take more care with what I am doing...

Well, let me give it a try.

1pm.

```
            op {
                fwd = if std < one then x else x / std
                bck = inl _ -> if std < one then one
                }
```

No this won't fly.

`if_` is giving me trouble. I do not want to CPS it, but I also do not have a good way of extracting the `bck`s from the if statements.

I give up. I'll do it roughly. It is not that important."