"4:20pm.
```
            k.block.iter (inl {item} ->
                inl adjoint = adjoint item
                adjoint .set (adjoint .get + out item .get)
                )
```

There is something wrong here. Should it really be `adjoint item` rather than `adjoint i`?

Ah, it really should be `item` after all. But what is going on with out. Should I not `get` it?

```
Error trace on line: 291, column: 13 in file "CudaAD".
            print_static out.unwrap
            ^
Error trace on line: 317, column: 36 in file "CudaAD".
                add_adjoint a (inl item -> bck.a (primal a item .get) (primal b item .get) (primal out item .get) * get_adjoint item out)
                                   ^
Cannot find a member named get inside the module.
```

Ah, `out` is a function. Ok.

4:25pm.

```
Running learning1.
Time for parse: 00:00:00.7786532
Time for prepass: 00:00:00.1839200
Time for peval was: 00:00:01.3914259

Process is terminated due to StackOverflowException.
Press any key to continue . . .
```

```
inl er = k.block.uter (+) out.adjoint
```

This happened as soon as I fixed this. This will take a bit to track down.

...Hmmm, it is not in `init_seq`.

4:30pm.

```
    inl relu_ln = seq float <| inl k x ->
            open CudaAD
            open Seq k
            succ x
```

Even this gives a SO error."