I am annoyed. I've backtracked through all the work I did in the Learning module and I still cannot locate the SO leak.
I can't believe the day is almost already over. I never thought it would take this long to do this, but I knew that it in situations like this it takes longer than one expects. I did well to leave it yesterday when I did.

One thing that I forgot to do is make a test for `init_seq`.

How about I do that now?