Ok, enough. I am done with layer norm. I am curious how it will do when I use them with RNNs and KFAC, but I am happy for the time being.
I did rather well today.

I definitely did a good bit. Amazing that it worked the first time when I ran it. Maybe I am getting better at this programming thing. Damn, I am tired right now. I am so drained that I actually feel sleepy. Zombieland ep 5 is out...

I think that making the norm functions `Op`s and taking the opportunity to make a firm distinction between `Activation`s and `Op`s is a good one. Making everything monadic is not really an option, but doing something as simple as imposing a convention on the two different modules rather than having some Activations be binded and some not is a great design choice.

It is unfortunate that seq functions require a return type explicitly, but I'll deal with that if it becomes an issue. I do not think it will.

I am really more worried about how I am going to do reversible operations. It will be hard, no doubt about it. But I will leave that for the future me. I am sure he will be able to handle it.

Tomorrow I will scrap those useless plastic RNNs and finally make the real one. That one will definitely be the next step in the evolution of artificial neural networks. It is the fundamental algorithmic step to everything that is going to come. Unless I take the time to do it right, it is impossible that I will reach any of my goals.