https://www.reddit.com/r/ProgrammingLanguages/comments/9tax2s/november_18_monthly_what_are_you_working_on_how/e8who53/?st=jnzs1rl3&sh=3d69a848
"I've done a lot of design work for the ML library. Starting with redesigning tensors and adding tree views, I've redesigned the initializers for the NN layers and apart from the wavefront iteration which was done over six months ago, implemented the entirety of the optimization suite from the old [Cuda blog](https://devblogs.nvidia.com/optimizing-recurrent-neural-networks-cudnn-5/) post.

Tree views are extremely convenient for doing data structure related optimizations by hand. What they are is very simple to describe - ordinary tensors in most languages are just multidimensional arrays, but in Spiral now it is possible to give the dimensions names.

Rather than have a tensor of size `100,150` one describe the dimensions as `{input=50; state=50}, {plastic=100; static=50}`. Once that is done, one can index into such a tensor like `tns ({input=()}, {static=()})` to get back a subtensor of size `50,50`. Alternatively, one can also use the tree based indexing `tns .view {input={static=()}}`. Of course this can be used diretly in Cuda kernels.

Right now I am using this to great effect to centralize all the weight matrices in a single tensor which is important not just performance and memory usage, but also making it easy to use the KFAC optimizer.

I've used it implement `segmented_map` which is like a map and concatenation operation rolled into one. Without tree views it would be impossible to initialize the tensor in the first place.

In the LSTM for example, I use it to [split the outputs](https://github.com/mrakgr/The-Spiral-Language/blob/92194ddd7afe0db09f8302dc2a9eb4791e339b77/Learning/LearningModule.fs#L1712) of the matrix multiply before passing them into the [fused activation](https://github.com/mrakgr/The-Spiral-Language/blob/92194ddd7afe0db09f8302dc2a9eb4791e339b77/Learning/LearningModule.fs#L264) separately.

The LSTM implementation that is in the library right now is a rather high quality one - there is a mutable optimization that I've opted not to do that could speed it up by another 10%, but apart from that it is very near the optimal point of what is possible.

[Spiral](https://github.com/mrakgr/The-Spiral-Language) is an excellent example of a language that uses high level abstractions to derive performance. Right now, having to fix type errors after a coding session is a fairly vicious task, but I find that after the program compiles I rarely have to do much fixing at runtime.

The [C++ samples](https://github.com/parallel-forall/code-samples/tree/master/posts/rnn) from the cuDNN blog are both ugly, non-composable and potentially buggy in comparison. One reason why Cuda programming is so dangerous is due to having to deal with calculating offsets for tensors by hand.

In contrast, the abstractions that I made last month will make it trivial for me to centralize the weights into a single multiply for any arbitrary NN architecture.

That was the work of last month. Right now, I am getting ready to finally deal with plastic RNNs which I consider to be the next step in the evolution of artificial neural networks and have just finished bringing back [layer norm](https://github.com/mrakgr/The-Spiral-Language/blob/92194ddd7afe0db09f8302dc2a9eb4791e339b77/Learning/LearningModule.fs#L429). I did a lot of work to enable automatic differentiation in Cuda kernels directly last month - [here](https://github.com/mrakgr/The-Spiral-Language/blob/92194ddd7afe0db09f8302dc2a9eb4791e339b77/Learning/old%20modules/LearningModule.fs'#L171) is how layer norm looked months ago. Quite a difference! Apart from using relu+LN as an activation, what I intend to do with layer norm is use it to normalize the weights in the Hebbian learning update.

The last two weeks were one shock after another for me. First I was surprised that KFAC does nothing for the bias weights and then I've discovered that on the binary pattern test the identity initialization gets nearly perfect score. That was embarrassing because it turns out that for entirety of September I was just testing how good the various architectures were at learning the identity. Plastic RNNs even without any parameters make very natural associative memories.

Just very lately I've been struggling to make modulation work with them. It was very annoying trying to figure out what is wrong with the optimizer and the architecture, but at the end of all of that I've gotten a surge of insight that cuts through all those complexities. I now know what the ideal form of the plastic RNN that does modulation and eligibility traces implicitly is.

As it turns out it is the simplest possible idea imaginable that I just had to think of last. Though I understand why the two papers might have missed it - as it involves having a plastic RNN on top of a vanilla static RNN, that still means one has to deal with all the instabilities having to optimize vanilla RNNs. But the age where RNNs were difficult optimize is past - in my tests I've discovered that when combined with KFAC, vanilla RNNs are actually on par with LSTMs (with KFAC) while having 4x less parameters. Both are significantly better than without KFAC.

With those pieces set, I have everything I need to enable the next degree of learning in my RL agents which will be essential for attaining victory in the battles to come. I see Hebbian learning as absolutely crucial and cannot see a future where RNNs using it are not the dominant architecture. Today I will verify that my idea works, and after that unfortunately I will have quite a bit of work to actually make it efficient. I was not at all counting to having to work on [reversible computation](https://www.youtube.com/watch?v=rVmZTGeIwnc) in order to achieve that, but that's algorithms for you.

There are different kinds of errors, and I really hope that the errors that will accumulate due to reversing the plastic weight updates will not be the kind that have negative impact on performance. Injecting Gaussian noise into the gradients is not bad for performance, but errors due to numerical inaccuracy are not random. They depend on the use case. Well, NNs tend to be robust so it should be fine.

The progress on Spiral is slow, but definite. The current deep learning wave will eventually, but inevitably carve out a path to a completely new kind of programming and lead to a crystallization that ability into true power. At the start of this year, I was pessimistic about the progress made, but now I think that the majority of trully influential research in the field has in fact been made since 2015 - that was the year of the DQN Atari paper, the bath norm paper, the residual net paper, the KFAC paper. 2018 has had a raft of research into metalearning of which plastic NNs are a part of.

My view back in 2015 was that the algorithms were garbage and it was going to be hard to use them to win even in something as simple as poker. Hence I thought I had time to things at my own pace. Without that kind of slack I would never have dared trying to make a ML library and eventually a language for that library. With every year and month that passes the situation changes and should the present pace continue, I have no doubt that ML is going to hit and exceed human level sometime in the next decade. Technology is the one true hope, and the work I am doing is my way of demonstrating my allegiance. So that my insight, my time and all of my power might enter a loop and come back to me stronger than when I'd last fired them."