"10:10am.
```
The learning rate is 2 ** -12
Starting timing for: iteration 0
The time was 00:00:06.3056697 for: iteration 0
Training: 2.90688475552837
The learning rate is 2 ** -12
Starting timing for: iteration 1
The time was 00:00:06.2746209 for: iteration 1
Training: 2.64245221528279
The learning rate is 2 ** -12
Starting timing for: iteration 2
The time was 00:00:06.3102401 for: iteration 2
Training: 2.60015857860665
The learning rate is 2 ** -12
Starting timing for: iteration 3
The time was 00:00:06.2558187 for: iteration 3
Training: 2.57440101125908
The learning rate is 2 ** -12
Starting timing for: iteration 4
The time was 00:00:06.3088611 for: iteration 4
Training: 2.55922267713365
```

It is better, but still not as good as feedforward. There is definitely something wrong with LN."

Let me resume. I needed a breather. Let me make the MIRNN. I need one more piece of evidence to make sure that LN is wrong.