I finally fixed all the type errors and got it to run. The MIRNN revealed a few type bugs I missed yesterday.
But I can be sure now - layer norm is incorrect. Well, that is hardly a surprise. Had it worked, it would have worked 100% with the vanilla RNN. I need to go through all the operations once more and make sure that they work. I probably got the gradients wrong for something.