"3:20pm.
```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:08.2678428 for: iteration 0
Training: 3.00185897701201
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:08.2416610 for: iteration 1
Training: 2.77824934773311
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:08.3001718 for: iteration 2
Training: 2.7172532977865
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:08.2451776 for: iteration 3
Training: 2.67970112104462
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:08.2263092 for: iteration 4
Training: 2.64926558076267
```

This `relu_ln`. I think it is finally time I ask myself one crucial question at this point. It is not too clear to me, but I think in the previous library I was using `ln + relu` rather than `relu + ln`. To be honest, I did not think there would be a distinction, but..

Did Mnist fool me into thinking I had to use `relu_ln` by necessity?

3:25pm.

```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:08.2161218 for: iteration 0
Training: 2.55665256511694
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:08.2366156 for: iteration 1
Training: 2.0240332518759
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:08.1793031 for: iteration 2
Training: 1.89091210763025
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:08.2191741 for: iteration 3
Training: 1.81517023843823
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:08.1851828 for: iteration 4
Training: 1.76115892756977
```

Oh, you son of a fucking bitch.

I did have the hunch that maybe I should use `ln_relu` for RNNs, but I ignored it earlier. I said to myself that `relu_ln` has to work.

Let me verify for sure whether it was relu+ln or ln+relu in the old library."