"3:45pm.
```
The learning rate is 2 ** -13
Starting timing for: iteration 0
The time was 00:00:05.7233478 for: iteration 0
Training: 3.0447242119733
The learning rate is 2 ** -13
Starting timing for: iteration 1
The time was 00:00:05.6583309 for: iteration 1
Training: 2.50614528906504
The learning rate is 2 ** -13
Starting timing for: iteration 2
The time was 00:00:05.6936235 for: iteration 2
Training: 2.32917928255355
The learning rate is 2 ** -13
Starting timing for: iteration 3
The time was 00:00:05.5242152 for: iteration 3
Training: 2.71546329293564
The learning rate is 2 ** -13
Starting timing for: iteration 4
The time was 00:00:05.6106155 for: iteration 4
Training: 5.53506827992661
```

This is the vanilla RNN with just `tanh` and no KFAC.

It seems that `ln_relu` is needed to stabilize it. There is a definitely a benefit to using `ln_relu` over `tanh` without KFAC.

`relu_ln` definitely works better than `ln_relu` on Mnist.

With KFAC there is no difference between `tanh` and `ln_relu`.

4:10pm. Had to take a break. I need to do the chores, but outside it is pouring. Ah, well, let me get to it.

4:30pm. Done with chores.

4:35pm. Damn, it is late. I do not feel like messing with the plastic RNN right now. But I have to. I must.

This issue today really sidetracked me massively, but the plastic RNN should not at all be difficult to implement.

How about I just sketch it out and then call it a day. It should not be hard.

On the bright side, I really did get it right the first time. Before I located the error, I even remember reasoning my way to the conclusion that there should not be any notable difference between `relu_ln` and `ln_relu` as far as RNNs are concerned. What a food I was.

Let me do it. I want to watch Index."