"5:15pm.
```
                inm {modulation out} =
                    inm data = segmented_init {dim=span,outer} {bias=const one; input=load input; state=load state}
                    inl data = Struct.map' (inl data -> data.basic) data
                    inm data = matmult_stream {weights with data}
                    wrap_split ((), {modulation=sublayer_size+size; out=size}) data
                inm out =
                    inm out' = matmult_stream {data=modulation; weight=H; streams=weights.streams}
                    tanh (out, out')
                inm H = wn_hebb {H modulation out}
                succ {out state={state=out; H}}
```

Let me do it like this. I almost named it input, but remembered at the last moment that this might change the order of dimensions. I'll just change `wn_hebb`. Let me modify it, put it as an activation and then I am done for the day. I am getting tired at this point.

5:20pm. Done. This should do it.

5:25pm. Beautiful. I removed those other useless plastic RNNs and cleaned up the library a little.

Let me paste it once more as it is so good.

```
    inl plastic_rnn size =
        {
        init = inl sublayer_size ->
            open Initializer.dual.TensorView
            inl outer =
                {
                static={bias=1; input=sublayer_size; state=size}
                plastic={input=sublayer_size; state=size}
                }
            inl inner =
                {
                static={modulation={input=sublayer_size; state=size} out=size}
                plastic=size
                }
            inl init =
                {
                bias={modulation={input=const zero; state=const zero}; out=const zero}
                input={modulation={input=identity; state=const zero}; out=identity}
                state={modulation={input=const zero; state=identity}; out=const zero}
                }
            {
            dsc =
                {
                weights = weight {init dim=outer.static,inner.static}
                dim = val {outer inner sublayer_size}
                }
            size
            }

        apply = inl {d with weights={weights dim={outer inner sublayer_size}} input} s ->
            inl span = primal input .span_outer
            assert (span = 1) "The differentiable plasticity layer supports only online learning for now."
            inl {state H} =
                match d with
                | {state} -> state
                | _ -> {
                    H=s.CudaTensor.zero {elem_type=float; dim=outer.plastic, inner.plastic}
                    state=s.CudaTensor.zero {elem_type=float; dim=span, size}
                    }

            inl apply =
                inm {modulation out} =
                    inm data = segmented_init {dim=span,outer} {bias=const one; input=load input; state=load state}
                    inl data = Struct.map' (inl data -> data.basic) data
                    inm data = matmult_stream {weights with data}
                    wrap_split ((), {modulation=sublayer_size+size; out=size}) data
                inm out =
                    inm out' = matmult_stream {data=modulation; weight=H; streams=weights.streams}
                    tanh (out, out')
                inm H = wn_hebb {H modulation out}
                succ {out state={state=out; H}}

            inl {out={out state} bck} = apply s
            {out state bck}
        optimize = Optimizer.kfac
        block = ()
        }
```

I do not care even if they are type errors lurking at the moment. This RNN is the pinnacle ML at this present point in time. It is the algorithm that will change the world. Once I make it my own, I will attain a truly strong power. I will have the ability to create agents that can not only glue themselves to the Nash equilibrium, but also learn to exploit weaker players. And in the games that will follow, these nets will give me the power and flexibility of agents with biologically inspired associative memories that are of importance for everything having to do with learning.

I do not see them as everything though - plasticity won't be enough to overcome catastrophic forgetting amongst other things, but once the right choice is made, future events will only move in the direction of bolstering the movement. The RL community will find some good things to help with exploration and sample complexity before long. Like I am doing now, I'll just pick and choose the best of the bunch. I believe in my ability to manage that.

Before that, I really need to make that agent so I can start making money otherwise I will be stuck with my GTX 970 forever. I need to move on from this - that FPGA Gameboy post that was posted recently on the RL sub is a direction I should be moving in.

I will definitely bring the plastic RNN saga to a conclusion in this November and start upscaling my activities in the real world."