"10:50am.
```
inl index_example' dim =
    Liple.foldr (inl x next l ->
        inl rec loop = function
            | {from near_to} -> next (dyn 0 :: l)
            | {} ->
                inl {k x} = module_foldr (inl k x _ -> {k x}) x ()
                {$k = loop x}
            | _ -> next (dyn 0 :: l)
        loop x
    ) dim (Tuple.rev >> Tuple.unwrap) ()
```

I've narrowed it down to here.

```
...
{k=type (type_lit (a)); x={input=lit 50i64; state=lit 50i64}}
```

Yeah, in the untested branch there is an infinite loop.

Ah, I get it. Shit.

```
            | {} as x ->
                inl {k x} = module_foldr (inl k x _ -> {k x}) x ()
                {$k = loop x}
```

This should take care of it.

10:55am.

```
inl data = Struct.map' (inl data -> data .view {a=()} .basic) data
```

It seems the view does not support partial application it seems. I will put it in.

11am.

```
            match i with
            | {} ->
                inl {k c i} = module_foldl (inl k s i -> {s with k i c=self+1}) {c=0} i
                assert (c = 1) "The number of branches in a view's index must be 1."
                inl x, s = f i (dim k)
                {$k=x}, s
            | _ ->
                (), i
```

Added the last branch. This should do it."