"11:15am.
```
Error trace on line: 58, column: 73 in file "View".
                                            | {from} -> Struct.map (inl x -> x-from) x, {s with near_to}
                                                                        ^
Error trace on line: 94, column: 11 in file "Core".
inl (-) a b = !Sub(a,b)
          ^
`is_numeric a && get_type a = get_type b` is false.
a={block=[]; from=lit 50i64; near_to=lit 100i64}, b=lit 0i64
```

Ah, I get it. Maybe I need to use `map'` instead of `map`. I probably did not account for the refactor.

11:15am.

```
Error trace on line: 6, column: 16 in file "Core".
inl error_type x = !ErrorType(x)
               ^
lit The tensor must be contiguous in order to be flattened.
```

Agh.

11:20am.

```
{bodies={ar=<function>; block=[]; offset=lit 0i64; size=[lit 101i64, lit 1i64]}; dim=[lit 1i64, lit 100i64]}
```

No this won't work. I thought it would be fine if it was in first place, but I was wrong.

Well, in the first place, for the `CudaBlas` function I do not need it to be contiguous. What i need is for the innermost dimension to not be padded.

I'll have to change how the library does its checking. I've been thinking about this issue for a while now. It is time to deal with it."