"4:30pm.
```
Error trace on line: 44, column: 9 in file "CudaAD".
        next {m with $k=x}
        ^
Error trace on line: 47, column: 32 in file "CudaAD".
inl unary_bind f a = bind (inl {a} -> f a) {a}
                               ^
Error trace on line: 107, column: 23 in file "CudaAD".
            fwd = inl x -> x * x
                      ^
Error trace on line: 96, column: 11 in file "Core".
inl (*) a b = !Mult(a,b)
          ^
`is_numeric a && get_type a = get_type b` is false.
a={a=<function>}, b={a=<function>}
```

Type errors as expected.

```
fwd = inl x -> x * x
```

This results in `{a=<function>}`.

4:35pm. What a strange error. I do not understand why is it happening.

4:40pm.

```
inl unary_op fwd bck a = op (inl {a} -> fwd a) {a = inl {a out} -> bck a out} {a}
```

I get it. It is because I am wrapping twice.

```
Training: 0.558679587916177
Testing: 0.369315119125904(8908/9984)
Training: 0.363312351557179
Testing: 0.320700551836919(9041/9984)
Training: 0.327145290075459
Testing: 0.2988061460738(9129/9984)
Training: 0.307631129446702
Testing: 0.285795027175202(9145/9984)
Training: 0.29486737276117
Testing: 0.277007831762043(9166/9984)
```

Ahhh, please just shot me now.

No, this is not because of the learning rate."