...Finally, I got it to compile. Let me see how it runs with a lr of 0.
```
At iteration 9 the cost and the sign accuracy are 72 and 0.64/1.
At iteration 19 the cost and the sign accuracy are 67.2 and 0.664/1.
At iteration 29 the cost and the sign accuracy are 104 and 0.48/1.
At iteration 39 the cost and the sign accuracy are 90.8 and 0.546/1.
```

Not that good. Let me let it run. I want to see if it will do the right thing. Probably the default LRs are way too large.

5:10pm.

```
The learning rate is 2 ** -10
At iteration 9 the cost and the sign accuracy are 72 and 0.64/1.
At iteration 19 the cost and the sign accuracy are 67.2 and 0.664/1.
At iteration 29 the cost and the sign accuracy are 104 and 0.48/1.
At iteration 39 the cost and the sign accuracy are 90.8 and 0.546/1.
At iteration 49 the cost and the sign accuracy are 79.6 and 0.602/1.
At iteration 59 the cost and the sign accuracy are 77.6 and 0.612/1.
```

No this is not good. I need smaller inits than the identity. How about 1/100th of an identity?