"5:55pm.
```
The learning rate is 2 ** -25
At iteration 9 the cost and the sign accuracy are 16.83619 and 0.994/1.
At iteration 19 the cost and the sign accuracy are 16.02943 and 1/1.
At iteration 29 the cost and the sign accuracy are 15.56203 and 0.994/1.
At iteration 39 the cost and the sign accuracy are 14.97722 and 0.996/1.
At iteration 49 the cost and the sign accuracy are 14.48278 and 0.992/1.
At iteration 59 the cost and the sign accuracy are 13.72867 and 0.982/1.
At iteration 69 the cost and the sign accuracy are 12.93931 and 0.994/1.
At iteration 79 the cost and the sign accuracy are 12.87379 and 0.996/1.
At iteration 89 the cost and the sign accuracy are 12.75586 and 0.994/1.
At iteration 99 the cost and the sign accuracy are 12.11186 and 0.998/1.
At iteration 109 the cost and the sign accuracy are 11.33875 and 0.986/1.
At iteration 119 the cost and the sign accuracy are 11.14428 and 0.994/1.
At iteration 129 the cost and the sign accuracy are 10.61888 and 0.992/1.
At iteration 139 the cost and the sign accuracy are 9.626036 and 0.988/1.
At iteration 149 the cost and the sign accuracy are 10.63905 and 0.996/1.
At iteration 159 the cost and the sign accuracy are 9.757484 and 0.99/1.
At iteration 169 the cost and the sign accuracy are 10.48168 and 0.982/1.
At iteration 179 the cost and the sign accuracy are 8.271495 and 0.998/1.
At iteration 189 the cost and the sign accuracy are 9.600921 and 0.972/1.
At iteration 199 the cost and the sign accuracy are 8.419868 and 0.992/1.
```

The error can be made to go down, but the learning rates needed are quite something.

6:40pm. Done with lunch. Let me call it a day here.

What I am going to do tomorrow is double down on layer norm. KFAC is working worse than SGD much like in the previous experiments.

I'll put `modulation` through `layer_norm >> tanh >> (*) 0.01f32` and I'll also switch to `ln_tanh` in the output. I'll drop the goal of trying to solve Binary Pattern and simply try to stabilize the training at all cost. Learning rates of -25 are around 2 ** 12 times too low for tastes.

I think that layer norm is in fact what I need to stabilize everything here. KFAC cannot really show what it is capable of if everything blows up on the first move, though there might be something more at play here as well.

Well, since this is my plan, I might as well try it now. I guess I can put in a bit of overtime."