"7:05pm. No, it is a failure. Hebbian learning is no joke - getting a handle on it is hard. I thought the modulation idea I had previously was the culprit, but it seems after I'd simplified it to the very limit, I still cannot do it.
I feel rather down because of this, but I need to investigate it more. I should not give up just yet. Even if Hebbian learning turns up empty, there is that other kind of memory from the Jimmy Ba paper that I could try.

There are some inexplicable things going on like when I init the modulation weights to zero the net blows up immediately with a Nan. Or how even after the tanh was maxed out, the outs kept growing. I need to make sure that weight norm is working properly.

7:15pm. Ah, how I want to continue. Now I feel sad and really want to make up for the loss, but this late is the worst time to start working again. Investigative programming needs to be done when fresh.

7:25pm. At worst, if I have to throw in the towel on this, I could always get rid of modulation and just use DP normally. At least that will be stable and I'd have an access to a RNN with memory.

7:30pm. Ah, let me at least figure out if weight norm is actually keeping the weights below the unit norm."