"8am.
```
At iteration 239 the cost and the sign accuracy are 27.07835 and 0.778/1.
At iteration 249 the cost and the sign accuracy are 25.67484 and 0.77/1.
At iteration 259 the cost and the sign accuracy are 19.91235 and 0.856/1.
At iteration 269 the cost and the sign accuracy are 50.05852 and 0.622/1.

Unhandled Exception: ManagedCuda.CudaException: ErrorAssert: A device-side assert triggered during kernel execution. The context cannot be used anymore, and must be destroyed. All existing device memory allocations from this context are invalid and must be reconstructed if the program is to continue using CUDA.
```

Nope, it is not any more stable at all. I need to be able to increase the LR from this level.

Let me try one of the ideas that I had last night.

```
(alpha * plastic + val one) * static >>= layer_norm >>= tanh
```

I meant this.

```
The learning rate is 2 ** -20
At iteration 1 the cost is NaN
The time was 00:00:00.0681388 for: Training
```

Yeah, no, it won't fly.

```
alpha * val 0.01f32 + static >>= layer_norm >>= tanh
```

Is `plastic` being state dependent a problem?

8:15am. I give up. This is not working. Basically, nothing can ever stabilize the plastic RNN in the presence of recurrent plastic connections.

Now that I think about it, the MI plastic RNN was miraculously stable in comparison even though it only ever needed to learn the identity.

Let me try the plastic RNN without the input plastic connections and then I will make my decision."