"9:05am.
```
-0.000000, 0.000000, -0.761590
-0.000000, 0.000000, -0.761590
nan, 0.000000, nan
nan, 0.000000, nan
```

So one of the outputs is Nan. Still, when does this happen?

9:10am.

```
            inl apply =
                inm data = segmented_init {dim=span,outer} {bias=const one; input=load input; state=load state}
                inl data = Struct.map' (inl data -> data.basic) data
                inm {alpha eta out} =
                    inm data = matmult_stream {weights with data}
                    succ (wrap_split ((), inner.static) data)
                //inm _ = print state
                //inm _ = print (Struct.map' (inl x -> x.basic) data)
                //inm _ = print modulation
                //inm _ = print out
                inm out =
                    //inm plastic = matmult_stream {data=state; weight={T=H}; streams=weights.streams; block=()}
                    //inm _ = print out'
                    //ln_tanh (alpha, plastic, out)
                    tanh out
                inm _ = print out
                inm H = wn_hebb {H eta out input=state}
                //inm _ = print weights.weight
                succ {out state={state=out; H}}
```

This has to be happening in wn_hebb. Probably the issue is not the multiplication by zero, but the fact that Nan gets passed into it somewhere. I should be careful of division.

Ah, something comes to mind. The derivative of a/b is db=-out/b.

If b is zero then...

Ah, so troublesome. I had not been expecting this at all.

I need to double check everything."