"10:50am. For my next project, how about I make it investigating the capabilities of CurveBall to deal with these kinds of models? Along the way I could study attractor nets for a bit.
For once I am not going to be implementing it myself. I'll go over into Python land and borrow an already existing PyTorch implementation.

10:55am. I'll end the morning session here, in defeat. I couldn't make this work. But I need to make a thorough investigation before I can conclude for sure.

...I've had experiences like this in the past before and they have usually led me to doing drastic things like making a language for example. This might very well the prelude to me delaying the entry into RL for a few months more so I can do higher order differentiation properly.

It might be assuming I can get CurveBall to work with plastic RNNs where KFAC failed me.

I was serious when I said that these nets are the future of ML. I will not compromise on this in the slightest.

Unless I get firm proof that what I am trying to do is impossible, I will move forward with this.

11am. No actually, wait. I should try KFAC on biases one more time.

Let me do that and then I will finally take a break. I've been at this for 4h now."