"1:20pm.
```
                    map (inl {alpha plastic static} ->
                        open CudaAD
                        open Op
                        alpha * plastic + static >>= tanh
                        ) {alpha plastic static=input}
```

The error is happening here. But why? It certainly was not happing in the previous net.

```
{bodies={ar=var (array_cuda_global (float32)); block=[]; offset=lit 0i64; size=[lit 50i64, lit 1i64]}; dim=[lit 1i64, lit 50i64]}
```

```
print_static in.plastic.primal.unwrap
```

No, there is nothing wrong here.

Then what?

1:30pm. Wow, I have absolutely no idea what the issue is.

```
    inl ln_tanh = seq float <| inl k {alpha plastic static} ->
        open CudaAD
        open Seq k
        open Op
        alpha * plastic + static >>= tanh
```

I managed to get it to compile using this instead.

I'll figure it out after I run the test.

1:35pm.

```
The learning rate is 2 ** -15
At iteration 9 the cost and the sign accuracy are 25.53054 and 0.998/1.
At iteration 19 the cost and the sign accuracy are 25.29893 and 1/1.
At iteration 29 the cost and the sign accuracy are 25.04969 and 0.992/1.
At iteration 39 the cost and the sign accuracy are 24.81659 and 1/1.
At iteration 49 the cost and the sign accuracy are 24.56994 and 0.992/1.
At iteration 59 the cost and the sign accuracy are 24.22754 and 0.996/1.
At iteration 69 the cost and the sign accuracy are 23.73145 and 1/1.
At iteration 79 the cost and the sign accuracy are 23.09989 and 1/1.
At iteration 89 the cost and the sign accuracy are 22.34789 and 0.994/1.
At iteration 99 the cost and the sign accuracy are 21.38993 and 0.998/1.
At iteration 109 the cost and the sign accuracy are 20.03851 and 0.99/1.
At iteration 119 the cost and the sign accuracy are 17.80331 and 0.996/1.
At iteration 129 the cost and the sign accuracy are 16.08945 and 1/1.
At iteration 139 the cost and the sign accuracy are 12.93467 and 1/1.
At iteration 149 the cost and the sign accuracy are 9.81106 and 1/1.
At iteration 159 the cost and the sign accuracy are 7.668978 and 1/1.
At iteration 169 the cost and the sign accuracy are 7.275251 and 0.974/1.
At iteration 179 the cost and the sign accuracy are 4.703811 and 0.994/1.
At iteration 189 the cost and the sign accuracy are 4.384847 and 0.994/1.
At iteration 199 the cost and the sign accuracy are 3.966154 and 0.994/1.
At iteration 209 the cost and the sign accuracy are 3.168599 and 0.988/1.
At iteration 219 the cost and the sign accuracy are 3.016284 and 0.992/1.
At iteration 229 the cost and the sign accuracy are 2.268794 and 0.996/1.
At iteration 239 the cost and the sign accuracy are 4.585472 and 0.98/1.
At iteration 249 the cost and the sign accuracy are 4.245306 and 0.972/1.
At iteration 259 the cost and the sign accuracy are 2.597123 and 0.996/1.
At iteration 269 the cost and the sign accuracy are 4.409456 and 0.976/1.
At iteration 279 the cost and the sign accuracy are 3.178078 and 0.98/1.
At iteration 289 the cost and the sign accuracy are 3.970897 and 0.978/1.
At iteration 299 the cost and the sign accuracy are 1.943707 and 0.996/1.
At iteration 309 the cost and the sign accuracy are 5.297452 and 0.972/1.
At iteration 319 the cost and the sign accuracy are 2.217508 and 0.994/1.
At iteration 329 the cost and the sign accuracy are 3.640543 and 0.988/1.
At iteration 339 the cost and the sign accuracy are 3.47885 and 0.978/1.
At iteration 349 the cost and the sign accuracy are 5.495466 and 0.968/1.
At iteration 359 the cost and the sign accuracy are 4.771472 and 0.976/1.
At iteration 369 the cost and the sign accuracy are 2.118162 and 0.996/1.
At iteration 379 the cost and the sign accuracy are 2.796269 and 0.984/1.
At iteration 389 the cost and the sign accuracy are 2.903352 and 0.986/1.
At iteration 399 the cost and the sign accuracy are 5.621906 and 0.966/1.
At iteration 409 the cost and the sign accuracy are 3.730412 and 0.978/1.
At iteration 419 the cost and the sign accuracy are 3.867807 and 0.982/1.
At iteration 429 the cost and the sign accuracy are 11.68221 and 0.926/1.
At iteration 439 the cost and the sign accuracy are 2.653002 and 0.994/1.
At iteration 449 the cost and the sign accuracy are 2.202761 and 0.994/1.
At iteration 459 the cost and the sign accuracy are 4.222591 and 0.982/1.
At iteration 469 the cost and the sign accuracy are 5.819108 and 0.974/1.
At iteration 479 the cost and the sign accuracy are 2.914144 and 0.99/1.
At iteration 489 the cost and the sign accuracy are 2.431782 and 0.988/1.
At iteration 499 the cost and the sign accuracy are 3.667733 and 0.982/1.
```

This would be great...except now I get this result without KFAC. Oggghhh...

1:40pm. I have no idea what is going on here. Right now I am in a state of complete confusion. First it did not work, and now it works.

Let me back it up here.

```
    inl plastic_rnn size =
        {
        init = inl sublayer_size ->
            open Initializer.dual.TensorView
            inl outer = {bias=1; input=sublayer_size; state=size}
            inl inner =
                {
                plastic=size
                bias={eta=1; alpha=1}
                }
            inl init =
                {
                alpha=const 0.01f32
                eta=const 0.01f32
                }
            {
            dsc =
                {
                bias = bias {init dim=1,inner.bias}
                streams = stream, stream
                dim = val {outer inner sublayer_size}
                }
            size
            }

        apply = inl {d with weights={bias streams dim={outer inner}} input} s ->
            inl span = primal input .span_outer
            assert (span = 1) "The differentiable plasticity layer supports only online learning for now."
            inl {state H} =
                match d with
                | {state} -> state
                | _ -> {
                    H=s.CudaTensor.zero_view {elem_type=float; dim=inner.plastic, outer} .basic
                    state=s.CudaTensor.zero {elem_type=float; dim=span, size}
                    }

            inl apply =
                inm token = Primitive.init {dim=span,1} (const one)
                inm {alpha eta} =
                    inm x = matmult_stream {bias with data=token}
                    succ (wrap_split ((), inner.bias) x)
                inm data = segmented_init {dim=span,outer} {bias=const one; input=load input; state=load state}
                inl data = Struct.map' (inl data -> data.basic) data

                inm out =
                    inm plastic = matmult_stream {data weight={T=H}; streams block=()}
                    ln_tanh {alpha plastic static=input}
                    //map CudaAD.Activation.hebb_tanh {alpha plastic static=input}
                inm H = wn_hebb {H eta out input=data}
                succ {out state={state=out; H}}

            inl {out={out state} bck} = apply s
            {out state bck}
        //optimize = Optimizer.kfac
        block = ()
        }
```

I still haven't figure out why using map here does not work. This is so crazy.

How about I figure out that type error and then I will figure out why optimization is going smoothly now. Could it be that there is something wrong with indexing?"